feat: config default write
The following commit accomplishes two objectives. First, it ensures the `run` folder and the `config.toml` file within it are ignored. This is to prevent changes to `config.toml` from being tracked in version control. Secondly, the `run` folder and the `config.toml` file will be automatically generated based on default settings if they do not already exist.