feat: arrow speeds and depletion
This uses correct arrow speeds and removes arrows from the inventory
when they're shot. This also includes some minor refactoring to the
inventory API that I thought would've been useful for implementing this
but ended up not using, but I left the changes in here since it'll
likely be useful later.