perf(hyperion): thread-local bump -> non-lazy (#30)
Before, when we were dealing with thread-local storage, we were lazily
initializing them. This is really inefficient if you're using an element
frequently because it requires a check each time. It can be substantial;
sometimes, it can even be twice as slow if done often. So instead, we're
now looking at the number of stages in the workflow and initializing one
buffer per stage.