feat: respawn (#702)
Current state:
- Works with and without enable_respawn_screen
- Restores player's health
- Change the player's team

Todo
- Respawn next to a teammate


When respawned, the client tries to teleport to its respawn location as
defined by PlayerSpawnPositionS2c, but the server prevents this as it
exceeds the maximum distance that the client is allowed to move in one
tick. This doesn't matter for the Tag event as we'll have to teleport
the player to another location anyway, but it could be an issue for
other events, so we should have a way to bypass this distance check when
it follows a respawn.

---------

Co-authored-by: Andrew Gazelka <andrew.gazelka@gmail.com>